requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
requests
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
python-dotenv
resend
